requirements.txt
black==24.4.2
click==8.1.7
dnspython==2.6.1
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pymongo==4.7.2
python-dotenv==1.0.1
